move safe_df_comparison to test_utilities
