+++ allensdk/test/brain_observatory/behavior/behavior_project_cache_data_model/test_behavior_project_cache.py
@@ -1,98 +1,9 @@
 import pytest
 import pandas as pd
 import logging
-import numpy as np
 import os
 
-
-def safe_df_comparison(expected: pd.DataFrame,
-                       obtained: pd.DataFrame,
-                       expect_identical_column_order: bool = False):
-    """
-    Compare two dataframes in a way that is agnostic to column order
-    and datatype of NULL values
-
-    Parameters
-    ----------
-    expected: pd.DataFrame
-
-    obtained: pd.DataFrame
-
-t    expect_identical_column_order: bool
-       If True, raise an error if columns are not
-       in the same order (default=False)
-
-    Raises
-    ------
-    RuntimeError
-       If:
-           - dataframes do not have the same columns
-           - dataframes do not have identical indexes
-           - dataframe columns do not have identical contents
-
-        When comparing the contents of dataframe columns,
-        the function:
-            - verifies that NULL values (whether None or NaN) are in
-              the same location
-            - loops over non-null values, casts arrays into lists, and
-              compares with ==
-    """
-    msg = ''
-    obtained_column_set = set(obtained.columns)
-    expected_column_set = set(expected.columns)
-
-    if obtained_column_set != expected_column_set:
-        msg += 'column mis-match\n'
-        msg += 'obtained columns\n'
-        msg += f'{obtained.columns}\n'
-        msg += 'expected columns\n'
-        msg += f'{expected.columns}\n'
-
-        missing_from_obtained = []
-        for c in expected.columns:
-            if c not in obtained.columns:
-                missing_from_obtained.append(c)
-        missing_from_expected = []
-        for c in obtained.columns:
-            if c not in expected.columns:
-                missing_from_expected.append(c)
-        msg += f'missing from obtained\n{missing_from_obtained}\n'
-        msg += f'missing from expected\n{missing_from_expected}\n'
-        raise RuntimeError(msg)
-
-    if not expected.index.equals(obtained.index):
-        msg += 'index mis-match\n'
-        msg += 'expected index\n'
-        msg += f'{expected.index}\n'
-        msg += 'obtained index\n'
-        msg += f'{obtained.index}\n'
-        raise RuntimeError(msg)
-
-    for col in expected.columns:
-        expected_null = expected[col].isnull()
-        obtained_null = obtained[col].isnull()
-        if not expected_null.equals(obtained_null):
-            msg += f'\n{col} not null at same point in '
-            msg += 'obtained and expected\n'
-            continue
-        expected_valid = expected[~expected_null]
-        obtained_valid = obtained[~obtained_null]
-        if not expected_valid.index.equals(obtained_valid.index):
-            msg += '\nindex mismatch in non-null when checking '
-            msg += f'{col}\n'
-        for index_val in expected_valid.index.values:
-            e = expected_valid.at[index_val, col]
-            o = obtained_valid.at[index_val, col]
-            if isinstance(e, np.ndarray):
-                e = list(e)
-            if isinstance(o, np.ndarray):
-                o = list(o)
-            if not e == o:
-                msg += f'\n{col}\n'
-                msg += f'expected: {e}\n'
-                msg += f'obtained: {o}\n'
-    if msg != '':
-        raise RuntimeError(msg)
+from allensdk.test_utilities.custom_comparators import safe_df_comparison
 
 
 @pytest.mark.parametrize("TempdirBehaviorCache, expected_ophys_session_table",
